check-files.py: initial commit
This utility attempts to identify any files required by ebuilds and verify that
they are available within ${FILESDIR}, and lists any files in ${FILESDIR} that
may not be required (such as the requiring ebuild being removed).

Results of this script should be verified before making changes to the
repository! Please report any issues found.

Signed-off-by: Sam Jorna <wraeth@wraeth.id.au>