Add script to parse bug-wranglers list
Add a script to parse the output of pybugz and attempt to identify the package
atom the bug relates to and list who should be set to assignee and which other
entities should be CC'd to the bug.

Note that this is highly experimental and should be verified before applying the
output to any bugs.