check-bugs.py: add filter for assignable address
This adds the option "--address" ('-a') which takes a single argument - an email
address. In the list of returned bugs, only those which have the specified email
address in the list of maintainers will be shown. This allows filtering results
by project - "only show unassigned bugs that should belong to proxy-maint".